Update sidebar text on outcome blur
Resolves #186
